Add check for a JSON element before attempting to match a string
This has previously caused massive performance issues, especially in
packs that frequently match non text elements. (see
https://spark.lucko.me/R6fJJ3zXOJ and the massive amounts of the trace
that is taken up by JsonSyntaxExceptions and MalformedJsonExceptions
that trace back to this line)

This patch first does a sanity check on the text element to see if it
ever could be a JSON element.